feat(Identity): Páginas front-end traduzidas

+++ RubyGameStoreWeb/Areas/User/Controllers/CarrinhoController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using RubyGameStore.Data.Repository.IRepository;
-using RubyGameStore.Helper;
+using RubyGameStore.Helper.StaticNames;
 using RubyGameStore.Models.Models;
 using RubyGameStore.Models.ViewModels;
 using Stripe.Checkout;
@@ -13,51 +13,63 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
     [Authorize]
     public class CarrinhoController : Controller
     {
-        private readonly IUnitOfWork unitOfWork;
+        private readonly IUnitOfWork _unitOfWork;
+        
         public CarrinhoVM CarrinhoVM { get; set; }
 
-        public CarrinhoController(IUnitOfWork _unitOfWork)
+        public CarrinhoController(IUnitOfWork unitOfWork)
         {
-            unitOfWork = _unitOfWork;
+            _unitOfWork = unitOfWork;
         }
 
         public IActionResult Index()
         {
             var claimsIdentity = (ClaimsIdentity)User.Identity;
             var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
-            CarrinhoVM = new CarrinhoVM()
+
+            CarrinhoVM = new()
             {
-                ListaCarrinho = unitOfWork.CarrinhoRepo.GetAll(c => c.UsuarioId == claim.Value, incluirPropriedades: "Produto,Produto.Plataforma"),
+                ListaCarrinho = _unitOfWork.CarrinhoRepo.GetAll(c => c.UsuarioId == claim.Value, incluirPropriedades: "Produto,Produto.Plataforma"),
                 PedidoCabecalho = new PedidoCabecalho()
             };
+
             foreach (var itemCarrinho in CarrinhoVM.ListaCarrinho)
             {
                 itemCarrinho.PrecoAtual = PrecoBaseadoNaQuantidade(itemCarrinho.Quantidade, itemCarrinho.Produto.Preco,
                     itemCarrinho.Produto.Preco50, itemCarrinho.Produto.Preco100);
                 CarrinhoVM.PedidoCabecalho.TotalPedido += itemCarrinho.Quantidade * itemCarrinho.PrecoAtual;
             }
+
             return View(CarrinhoVM);
         }
 
-        public IActionResult RevisarPedido()
+        public IActionResult CriarPedido()
         {
             var claimsIdentity = (ClaimsIdentity)User.Identity;
             var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
+            var usuario = _unitOfWork.UsuarioRepo.GetFirstOrDefault(u => u.Id == claim.Value);
 
             CarrinhoVM = new CarrinhoVM()
             {
-                ListaCarrinho = unitOfWork.CarrinhoRepo.GetAll(c => c.UsuarioId == claim.Value, incluirPropriedades: "Produto,Produto.Plataforma"),
+                ListaCarrinho = _unitOfWork.CarrinhoRepo.GetAll(c => c.UsuarioId == claim.Value, incluirPropriedades: "Produto,Produto.Plataforma"),
                 PedidoCabecalho = new PedidoCabecalho()
+                {
+                    UsuarioId = usuario.Id,
+                    NomeDestinatario = $"{usuario.NomeUsuario} {usuario.SobrenomeUsuario}",
+                    TelefoneContato = usuario.TelefoneContato,
+                    LogradouroEntrega = usuario.LogradouroUsuario,
+                    CidadeEntrega = usuario.CidadeUsuario,
+                    EstadoEntrega = usuario.EstadoUsuario,
+                    CEPEntrega = usuario.CEPUsuario
+                }
             };
 
-            CarrinhoVM.PedidoCabecalho.Usuario = unitOfWork.UsuarioRepo.GetFirstOrDefault(u => u.Id == claim.Value);
-
-            CarrinhoVM.PedidoCabecalho.NomeDestinatario = $"{CarrinhoVM.PedidoCabecalho.Usuario.NomeUsuario} {CarrinhoVM.PedidoCabecalho.Usuario.SobrenomeUsuario}";
-            CarrinhoVM.PedidoCabecalho.TelefoneContato = CarrinhoVM.PedidoCabecalho.Usuario.TelefoneContato;
-            CarrinhoVM.PedidoCabecalho.LogradouroEntrega = CarrinhoVM.PedidoCabecalho.Usuario.LogradouroUsuario;
-            CarrinhoVM.PedidoCabecalho.EstadoEntrega = CarrinhoVM.PedidoCabecalho.Usuario.EstadoUsuario;
-            CarrinhoVM.PedidoCabecalho.CidadeEntrega = CarrinhoVM.PedidoCabecalho.Usuario.CidadeUsuario;
-            CarrinhoVM.PedidoCabecalho.CEPEntrega = CarrinhoVM.PedidoCabecalho.Usuario.CEPUsuario;
+            CarrinhoVM.PedidoCabecalho.Usuario = _unitOfWork.UsuarioRepo.GetFirstOrDefault(u => u.Id == claim.Value);
+            if (RedirecionarCadastro(CarrinhoVM))
+            {
+                TempData["StatusMessage"] = "É necessário que você complete seu cadastro para poder realizar compras!";
+                return RedirectToPage("/Account/Manage/Index", new { area = "Identity" });
+            }            
 
             foreach (var itemCarrinho in CarrinhoVM.ListaCarrinho)
             {
@@ -65,25 +77,21 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
                     itemCarrinho.Produto.Preco50, itemCarrinho.Produto.Preco100);
                 CarrinhoVM.PedidoCabecalho.TotalPedido += itemCarrinho.Quantidade * itemCarrinho.PrecoAtual;
             }
-
             return View(CarrinhoVM);
         }
 
         [HttpPost]
-        [ActionName("RevisarPedido")]
         [ValidateAntiForgeryToken]
-        public IActionResult RevisarPedidoPOST(CarrinhoVM CarrinhoVM)
+        public IActionResult CriarPedido(CarrinhoVM CarrinhoVM)
         {
             var claimsIdentity = (ClaimsIdentity)User.Identity;
             var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
+            if (claim.Value != CarrinhoVM.PedidoCabecalho.UsuarioId) return RedirectToPage("/Account/AccessDenied", new { area = "Identity" });
 
-            CarrinhoVM.ListaCarrinho = unitOfWork.CarrinhoRepo.GetAll(c => c.UsuarioId == claim.Value, incluirPropriedades: "Produto,Produto.Plataforma");
-
-            CarrinhoVM.PedidoCabecalho.StatusPedido = StaticDetails.StatusPendente;
-            CarrinhoVM.PedidoCabecalho.StatusPagamento = StaticDetails.PagamentoPendente;
+            CarrinhoVM.ListaCarrinho = _unitOfWork.CarrinhoRepo.GetAll(c => c.UsuarioId == claim.Value, incluirPropriedades: "Produto,Produto.Plataforma");            
             CarrinhoVM.PedidoCabecalho.DataHoraPedido = DateTime.Now;
-            CarrinhoVM.PedidoCabecalho.UsuarioId = claim.Value;
-
+            CarrinhoVM.PedidoCabecalho.StatusPedido = Pedido.Pendente;
+            CarrinhoVM.PedidoCabecalho.StatusPagamento = Pagamento.Pendente;            
             foreach (var itemCarrinho in CarrinhoVM.ListaCarrinho)
             {
                 itemCarrinho.PrecoAtual = PrecoBaseadoNaQuantidade(itemCarrinho.Quantidade, itemCarrinho.Produto.Preco,
@@ -91,8 +99,8 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
                 CarrinhoVM.PedidoCabecalho.TotalPedido += itemCarrinho.Quantidade * itemCarrinho.PrecoAtual;
             }
 
-            unitOfWork.PedidoCabecalhoRepo.Add(CarrinhoVM.PedidoCabecalho);
-            unitOfWork.Save();
+            _unitOfWork.PedidoCabecalhoRepo.Add(CarrinhoVM.PedidoCabecalho);
+            _unitOfWork.Save();
 
             foreach (var itemCarrinho in CarrinhoVM.ListaCarrinho)
             {
@@ -103,17 +111,18 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
                     Preco = itemCarrinho.PrecoAtual,
                     Quantidade = itemCarrinho.Quantidade
                 };
-                unitOfWork.PedidoDetalhesRepo.Add(pedidoDetalhes);
-                unitOfWork.Save();
+                _unitOfWork.PedidoDetalhesRepo.Add(pedidoDetalhes);
+                _unitOfWork.Save();
             }
 
             if (User.IsInRole("Empresa"))
             {
                 return RedirectToAction("Confirmado", "Carrinho", new { id = CarrinhoVM.PedidoCabecalho.Id });
             }
+
+            //Stripe
             else
             {
-                //Stripe
                 var domain = "https://localhost:7213/";
                 var options = new SessionCreateOptions
                 {
@@ -143,24 +152,29 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
 
                 var service = new SessionService();
                 Session session = service.Create(options);
-                unitOfWork.PedidoCabecalhoRepo.AtualizarStatusStripe(CarrinhoVM.PedidoCabecalho.Id, session.Id, session.PaymentIntentId);
-                unitOfWork.Save();
+                _unitOfWork.PedidoCabecalhoRepo.AtualizarStatusStripe(CarrinhoVM.PedidoCabecalho.Id, session.Id, session.PaymentIntentId);
+                _unitOfWork.Save();
                 Response.Headers.Add("Location", session.Url);
+
                 return new StatusCodeResult(303);
             }
-
         }
 
         public IActionResult Confirmado(int id)
         {
-            var pedidoCabecalho = unitOfWork.PedidoCabecalhoRepo.GetFirstOrDefault(p => p.Id == id);
-            var carrinho = unitOfWork.CarrinhoRepo.GetAll(p => p.UsuarioId == pedidoCabecalho.UsuarioId).ToList();
+            var pedidoCabecalho = _unitOfWork.PedidoCabecalhoRepo.GetFirstOrDefault(p => p.Id == id);
+            var listaCarrinho = _unitOfWork.CarrinhoRepo.GetAll(p => p.UsuarioId == pedidoCabecalho.UsuarioId).ToList();
+
+            var claimsIdentity = new ClaimsIdentity(User.Claims);
+            var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
+
+            if(pedidoCabecalho.UsuarioId != claim.Value) return RedirectToPage("/Account/AccessDenied", new { area = "Identity" });
 
             if (User.IsInRole("Empresa"))
             {
-                unitOfWork.PedidoCabecalhoRepo.AtualizarStatus(pedidoCabecalho.Id, StaticDetails.StatusAprovado, StaticDetails.PagamentoConvenio);
-                unitOfWork.CarrinhoRepo.RemoveRange(carrinho);
-                unitOfWork.Save();
+                _unitOfWork.PedidoCabecalhoRepo.AtualizarStatus(pedidoCabecalho.Id, Pedido.Aprovado, Pagamento.Empresarial);
+                _unitOfWork.CarrinhoRepo.RemoveRange(listaCarrinho);
+                _unitOfWork.Save();
             }
             else
             {
@@ -169,9 +183,9 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
 
                 if (session.PaymentStatus.ToLower() == "paid")
                 {
-                    unitOfWork.PedidoCabecalhoRepo.AtualizarStatus(pedidoCabecalho.Id, StaticDetails.StatusAprovado, StaticDetails.PagamentoAprovado);
-                    unitOfWork.CarrinhoRepo.RemoveRange(carrinho);
-                    unitOfWork.Save();
+                    _unitOfWork.PedidoCabecalhoRepo.AtualizarStatus(pedidoCabecalho.Id, Pedido.Aprovado, Pagamento.Aprovado);
+                    _unitOfWork.CarrinhoRepo.RemoveRange(listaCarrinho);
+                    _unitOfWork.Save();
                 }
             }
 
@@ -180,26 +194,26 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
 
         public IActionResult Incrementar(int carrinhoId)
         {
-            var carrinhoDB = unitOfWork.CarrinhoRepo.GetFirstOrDefault(c => c.Id == carrinhoId);
-            unitOfWork.CarrinhoRepo.AdicionarAoCarrinho(carrinhoDB, 1);
-            unitOfWork.Save();
+            var carrinhoDB = _unitOfWork.CarrinhoRepo.GetFirstOrDefault(c => c.Id == carrinhoId);
+            _unitOfWork.CarrinhoRepo.AdicionarAoCarrinho(carrinhoDB, 1);
+            _unitOfWork.Save();
             return RedirectToAction("Index");
         }
 
         public IActionResult Decrementar(int carrinhoId)
         {
-            var carrinhoDB = unitOfWork.CarrinhoRepo.GetFirstOrDefault(c => c.Id == carrinhoId);
-            if (carrinhoDB.Quantidade == 1) unitOfWork.CarrinhoRepo.Remove(carrinhoDB);
-            else unitOfWork.CarrinhoRepo.RemoverDoCarrinho(carrinhoDB, 1);
-            unitOfWork.Save();
+            var carrinhoDB = _unitOfWork.CarrinhoRepo.GetFirstOrDefault(c => c.Id == carrinhoId);
+            if (carrinhoDB.Quantidade == 1) _unitOfWork.CarrinhoRepo.Remove(carrinhoDB);
+            else _unitOfWork.CarrinhoRepo.RemoverDoCarrinho(carrinhoDB, 1);
+            _unitOfWork.Save();
             return RedirectToAction("Index");
         }
 
         public IActionResult Remover(int carrinhoId)
         {
-            var carrinhoDB = unitOfWork.CarrinhoRepo.GetFirstOrDefault(c => c.Id == carrinhoId);
-            unitOfWork.CarrinhoRepo.Remove(carrinhoDB);
-            unitOfWork.Save();
+            var carrinhoDB = _unitOfWork.CarrinhoRepo.GetFirstOrDefault(c => c.Id == carrinhoId);
+            _unitOfWork.CarrinhoRepo.Remove(carrinhoDB);
+            _unitOfWork.Save();
             return RedirectToAction("Index");
         }
 
@@ -208,6 +222,16 @@ namespace RubyGameStoreWeb.Areas.User.Controllers
             if (quantidade > 100) return preco100;
             else if (quantidade >= 50 && quantidade < 100) return preco50;
             else return preco;
+        }
+
+        public bool RedirecionarCadastro(CarrinhoVM vm)
+        {
+            if (vm.PedidoCabecalho.Usuario.TelefoneContato == null ||
+                vm.PedidoCabecalho.Usuario.LogradouroUsuario == null ||
+                vm.PedidoCabecalho.Usuario.CidadeUsuario == null ||
+                vm.PedidoCabecalho.Usuario.EstadoUsuario == null ||
+                vm.PedidoCabecalho.Usuario.CEPUsuario == null) return true;
+            else return false;
         }
     }
 }
